ElevateChat&AIApplications/code/backend/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
langchain
openai
openai[datalib]
azure-storage-blob
azure-identity
azure-core
unstructured 
tiktoken
#pre release https://pypi.org/project/azure-search-documents/#history
azure-search-documents==11.4.0b6
